All the IGeometry subclasses are updated to use size_t instead of int64_t
When asking for the number of items/elements in a geometry the type should be a size_t
since those values are natively stored as size_t in the underlying classes.

+++ Source/SIMPLib/Geometry/EdgeGeom.cpp
@@ -147,7 +147,7 @@ EdgeGeom::~EdgeGeom() = default;
 // -----------------------------------------------------------------------------
 //
 // -----------------------------------------------------------------------------
-EdgeGeom::Pointer EdgeGeom::CreateGeometry(int64_t numEdges, SharedVertexList::Pointer vertices, const QString& name, bool allocate)
+EdgeGeom::Pointer EdgeGeom::CreateGeometry(size_t numEdges, SharedVertexList::Pointer vertices, const QString& name, bool allocate)
 {
   if(name.isEmpty())
   {
