+++ Source/SIMPLib/Geometry/TriangleGeom.cpp
@@ -149,7 +149,7 @@ TriangleGeom::~TriangleGeom() = default;
 // -----------------------------------------------------------------------------
 //
 // -----------------------------------------------------------------------------
-TriangleGeom::Pointer TriangleGeom::CreateGeometry(int64_t numTriangles, SharedVertexList::Pointer vertices, const QString& name, bool allocate)
+TriangleGeom::Pointer TriangleGeom::CreateGeometry(size_t numTriangles, SharedVertexList::Pointer vertices, const QString& name, bool allocate)
 {
   if(name.isEmpty())
   {
